Fix: unlinked nodes may be freed before document (leak)
The document only has weak references to its nodes, so a node may be
collected before the document... but if the node was unlinked the
document's finalizer wouldn't know about the node anymore, and it leaked
the libxml node.

Fixes the issue by keeping a list of unlinked libxml nodes.