Allow expressions after non-trailing `if` `MacroExpression` (#15917)
The following is already valid: (this is not a `MacroIf` node, but a `MacroExpression` node whose first sub-expression is an `If`, because the `if` is closed within the braces)

```crystal
{%
  2 if 1
  3
%}
```

This PR adds support for the equivalent with a non-trailing `if`:

```crystal
{%
  if 1
    2
  end
  3
%}
```

In fact, macro interpolation already transforms the former into the latter; these previously broken interpolations will now work.

`{% unless ...; end; ... %}` is also handled in the same way.

`{{ ... }}` expects a single expression and never an expression list in the first place, so that form is unaffected.