Fix: XML::TextReader must keep IO object reference (#15933)
Running the std xml specs could sometimes lead to a segfault. The problem is that the IO object passed to XML::TextReader can be collected by the GC before the XML::TextReader is finalized, leading to a segfault when libxml trying to close the input buffer, hence close the IO object, and :boom:

Also harmonizes the callbacks to use `->method` with the rest of the libxml library.