Fix: remove the minimum in Fiber::ExecutionContext::Parallel
The context should be in control of the overall parallelism of the
context. In light of the incoming thread pool it makes more sense for
threads to be returned to the pool, so they can increase the parallelism
of contexts that need it, without starting more and more system
threads.