Print deprecation warning on types and aliases
We could already deprecate a type and have the documentation report the
deprecation. We now also print a deprecation warning during compilation
when we try to access the type, when calling a class method on the type
(no need to deprecate every method) but also when extending a type
during a class (or struct) definition.

There are more cases that haven't been implemented, yet (include,
extend, instance methods, ...).

We can also now deprecate an alias, which prints a deprecation warning
and documents the deprecation.