Rename execution contexts + improve their docs (#15936)
Execution contexts are much more about "how fibers run", either concurrently or in parallel or by themselves (the intent), and have little to do with "threads" that are only the technical mean to achieve parallelism.

A great example is that the isolated context doesn't have the word thread in its name, despite being the only context that guarantees the fiber will always run on the same thread (the fiber owns the thread until it terminates). 

This PR renames:

- `SingleThreaded` to `Concurrent`
- `MultiThreaded` to `Parallel`

It also refines the documentation of the different types to also insist on the intent, what it means, and what are the actual guarantees given by each context.

Co-authored-by: Johannes Müller <straightshoota@gmail.com>
Co-authored-by: sbsoftware <stefan@sbsoftware.de>