Make max_size a macro variable
This commit reuses the same calculation to avoid the mental overhead of
figuring out what, if anything, was different between the two. Now it's
obvious that they are the same calculation.

There is no runtime performance difference. LLVM was likely already
inlining `max_size` as a constant because we were assigning a numeric
literal to the variable, resulting in what we are currently doing.

Co-authored-by: Johannes Müller <straightshoota@gmail.com>
Co-authored-by: Lars Olsson <lasso@lassoweb.se>