Load `Location.local` from symlink name (#16002)
Previously, `Time::Location.local` on Unix systems would always have a generic name, `Local` (or `UTC` if `/etc/localtime` is missing) because the zoneinfo described in `/etc/localtime` does not have an obvious name attached.

`/etc/localtime` is typically a symlink into the zoneinfo database. With this change, we read the symlink and defer the name of the location from the path inside the zoneinfo root directory. We do not actually parse the target file and instead load the location by name.

Originally suggested in https://forum.crystal-lang.org/t/how-do-you-get-the-location-name-of-the-local-timezone/8191/3?u=straight-shoota, with alterations from https://github.com/crystal-lang/crystal/pull/15959#issuecomment-3035370559 (same as jiff).