Remove minimum in `Fiber::ExecutionContext::Parallel` (#15946)
Having a minimal number of threads felt normal, but starting a parallel context with a minimum starts that number of schedulers and threads, but there's nothing to do, so the schedulers park the threads almost immediately, and even before we could spawn a fiber into the context.

The same happens when the context has no fibers to run: the schedulers park, and the parallelism returns to zero.

_The actual parallelism will always reach zero instead of the configured minimum._

In practice, the range is `0..max` not `min..max`: every parallel context starts at zero, increases to one when a first fiber is pushed, then the context will decide to increase the parallelism (assign more threads to run the schedulers) and decrease the parallelism based on the context current needs, up to the maximum parallelism, that may or may not be reached.

NOTE: the range is actually `0` on creation, then `1..max` because we keep a thread to run the context's event loop (that might change in the future, if the context knows it has zero fibers).