Optimize `Parallel::Scheduler#quick_dequeue?` when capacity is 1 (#15961)
Brings an optimization from the `Concurrent` context to the `Parallel` context when the parallelism is set to 1: we can spare a switch to the scheduler's run loop fiber to (not so) quickly check the global queue and event loop.

With this, there's no noticeable speed difference between `Concurrent` and `Parallel` with a single scheduler in a few benchmarks, such as a swarm of fibers or a HTTP server.