Add `REUSE.toml` (#15992)
`REUSE.toml` specifies the licenses associated with different parts of this repository according to the [reuse specification](https://reuse.software/).

Also specifies the SPDX identifier as `Apache-2.0 WITH Swift-exception` instead of only `Apache-2.0`.

This article is a good introduction of the purpose and operation: https://fedoramagazine.org/beginners-guide-for-open-source-developers-for-software-licensing-with-fsfe-reuse/

There are no changes to the licenses or licensing terms. This patch js just a formal declaration of which licenses apply to which parts of the code base.