Lazily instantiate the event loop of isolated execution contexts (#16063)
An isolated context might not need an event loop and certainly doesn't need one upfront. We can delay its instantiation to when we need it.

The fiber is isolated, so it doesn't need to bother with concurrency or parallelism issues, however the context must support two mechanisms to suspend the fiber/thread: wait on the event loop (when instantiated) and wait on a condition variable otherwise.