Allow to set a custom base href in docs' 404 page
Since the 404 page is served at different (sub)paths in a website,
relative references to assets (CSS and JS files) don't work for
subdirectories.

We now add a `--404-base-href` flag to the docs generator to explicitly
set that path.

Adding that `<base href />` tag in other pages makes hash refs
(`#links`) relative to that path, so that would break every other page.