Fix return type of `system_close_on_exec=` on Windows (#16095)
`#system_close_on_exec=` (and thus `#close_on_exec=`) returns `Nil` on Windows. But #15698 added a type-restriction to `Bool`. We didn't notice this discrepancy before, because this method was never called on Windows until we enabled the specs for that in #14716.

Changing the return value to `false` on Windows should be the correct behaviour. If the parameter is `true`, the method raises.

This is a classic integration issue. It could've been avoided by testing each change against current master directly before merging it. But luckily they are pretty rare, so there's no urgent need to do something about it.