Resolve types when guessing return type from class method overloads (#16118)
In the following snippet, the compiler is able to infer `@foo`'s type to be `Foo`, since both overloads of `Foo.foo` have a return type of `Foo`:

```crystal
class Foo
  def self.foo(x : Int32) : Foo
    new
  end

  def self.foo(x : String) : Foo
    new
  end
end

class Bar
  def initialize
    @foo = Foo.foo(1)
  end
end

Bar.new
```

Changing either return type to `::Foo` or `self` breaks the type guesser, because currently all the type names are required to be identical in spelling. Those names are now resolved so that different spellings can be used (an example in the standard library is `Log::Metadata.build`, whose overloads return `Metadata` and `self` respectively).