Add `#has_any_args?` method for `ASTNode`s (#16115)
Several `ASTNode` implementations have a concept of arguments: `Call`, `Def`, `FunDef`, `Macro`, `ProcPointer`, `Block`, `Annotation`.
Arguments are often separated into different properties, such as `args`, `named_args` and `block_args` on `Call`. In order to know whether the node has any args, we need to check if either of them is present. This is a common pattern throughout the compiler code base. The tricky thing is to remember, for each node type, which properties need to be checked. In some cases we're not checking for all conditions. Whether this is a relevant omission, depends on the exact circumstance.

This patch introduces the `#has_any_args?` method on all node types with arguments. Even those where it's really just an `args` property (`Block` and `ProcPointer`). This makes it very convenient to check for the absence of any arguments without having to worry about the node type's data layout and potentially forgetting something.

Co-authored-by: Sijawusz Pur Rahnama <sija@sija.pl>