Disable flaky spec for `Fiber::ExecutionContext::GlobalQueue` on macOS (#16146)
The failure happens on both x86_64 and AArch64 and happens too often on CI. An attempt to fix it didn't work (#15630). I assume macOS threads and sleeps are doing something different than the spec expects 🤷

It never happens on other platforms which should be enough to stress test the thread safety of the feature.