Default execution context is now parallel (MT:1) (#16136)
The concurrent context doesn't exhibit any performance improvement over the parallel context (see #16135). So we may as well use a parallel context as default context, and allow resizing it for more parallelism (see #15956).

The default configuration is without parallelism (MT:1), thus the execution context remains concurrent only, which is backward compatible (nothing changes).

Resizing the context to enable parallelism will be at the discretion of the application: it can keep using the `CRYSTAL_WORKERS` environment variable, and/or fallback to the number of CPUs or use a hardcoded value, and/or provide a command line argument, or anything else.

Co-authored-by: Johannes Müller <straightshoota@gmail.com>