Always use the custom table so references won't be collected by the GC
The original implementation required to keep a reference to any GC
memory stored in Thread::LocalStorage when using the TSS or FLS
fallbacks because the GC can't scan the tables and will consider the
memory to be unreachable and collect it (oops).

This refactor always uses the custom table and only stores the pointer
to the table in TSS/FLS which is safe because a pointer to the table is
always kept in reachable memory.