Add Thread::LocalStorage
Relies on a macro to inject typed instance variables into the container,
and optional destructors as class variables. The overall design is very
simple and accessing a thread local is basically dereferencing the
actual thread local and accessing an ivar.

One drawback is that the compiler doesn't drop unused ivars, so as soon
as an executable needs a thread local, it will allocate the whole type,
even though the program only uses a few of them. Since we don't intent
to have lots of thread locals, this ain't much of an issue.

Another drawback is that destructors will always be registered, whatever
if we actually need the thread local, which may lead to always link
libpcre2 for example, because the destructors will link the external
free symbols, which can be problematic. Maybe weak symbols could help?