Revert "Add `Thread::Local(T)` (#15616)" (#16179)
This reverts commit 8fda88f0cfa13265bcd8a72da9e2d362a13ce9dd.

It's an internal type, we don't plan to use it anymore, and #16173 is
meant to replace it.

It's main feature was the ability to work on every target (unlike the current `@[ThreadLocal]` annotation) and to register destructors... but the annotation can be fixed using the EmulatedTLS LLVM option, and the destructor is crashing on x86_64-darwin when mixing pthreads, unwind and pcre2.