Add Crystal::FdLock to count system fd references
This patch implements a reference counted lock to protect IO objects
that depend on a reusable system fd (IO::FileDescriptor, File and
Socket) to protect them against thread safety issues around close:

- Thread 1 wants to read from fd 123;
- The OS preempts Thread 1;
- Thread 2 closes fd 123;
- Thread 2 opens something else and the OS reuses fd 123;
- The OS resumes Thread 1;
- Thread 1 reads from the reused fd 123!!!

The issue arises for any operation that would mutate the fd: write,
fchown, ftruncate, setsockopt, ... as they risk affecting a reused fd
instead of the expected one.

NOTE: The patch provides two implementations of Crystal::FdLock, one is
thread safe for MT, while the other is thread unsafe because the
interpreter segfaults when using the thread safe implementation, despite
the interpreted code being single threaded (it has something to do with
the refcount and resuming the closing fiber).