PCRE2: use thread local for jit stack and match data (#16175)
We can reuse the same JIT stack and matchdata for all instances. There's no need for a specific match data per `Regex`: We just allocate one and make sure it's not used by other threads.

The drawback is that we must allocate each matchdata with a maximum number of ovectors (65535). That might increase memory usage, though I failed to notice it in practice.
Maybe not allocating separate memory for every regular expression is helping?