`Log::Metadata` should put parent entries first on extend (like `Hash#merge`) (#16098)
`Log::Metadata` wrote its entries in the beginning of the allocated memory for all entries. When extending, parent entries were prepended to the memory.

With this change, entries are written to the end of allocated memory to make it possible to write parent entries first when extending the `Metadata` instance. `#defrag` moves the local entries to the end of the parent entries or overwrites any existing parent entry for the current key.