EC: register schedulers to event loop (extract PR)
Some event loops need to be aware of the execution context schedulers,
to allocate and cleanup resources as schedulers are started and
stopped.

For example the io_uring evloop will need to keep a ring per thread, yet
can reuse the main ring for the first scheduler (index 0), and also
needs to keep a list of all the active rings.

- Add `Crystal::EventLoop#initialize(parallelism)`.
- Add `Crystal::EventLoop#register(scheduler, index)`.
- Add `Crystal::EventLoop#unregister(scheduler)`.