Deprecate single-letter macro fresh variables with indices (#16267)
Crystal supports 6 sigils which modify the meaning of string literals. Curly braces may be used instead of the usual double quotation marks for those strings:

```crystal
%i{1}
%q{2}
%Q{3}
%r{4}
%x{5}
%w{6}
```

The same syntax is used for fresh variables inside macros:

```crystal
macro foo
  {% for i in 1..6 %}
    %b{i} = {{ i }}
  {% end %}
  {% for i in 1..6 %}
    p %b{i}
  {% end %}
end

foo
```

It is a syntax error here to use a fresh variable name that coincides with one of the sigils, because the sigil has precedence and assignment to a string literal is never allowed. This is problematic since adding _any_ new sigil to the language, e.g. `%b` for #2886, immediately introduces a breaking change by making those macros no longer valid.

I believe the language should be more future-proof here, unless we can say for sure that new sigils must not be introduced.

Co-authored-by: fn ⌃ ⌥ <70830482+FnControlOption@users.noreply.github.com>