[CI] Add lint workflow running `pre-commit` (#16275)
It adds a new _Lint_  workflow which runs `pre-commit` in CI. It only tests files that have changes compared to the base ref of a PR, or `master`.
Workflow dispatch allows running `pre-commit` against all files in the repo.

We already have a couple of individual CI jobs that run some of these linter: `lint-actionlint`, `lint-shellcheck`, `lint_spellcheck`. The new setup replaces them. But I'm keeping them around for a while to ensure the new workflow functions properly.