Add specs for `Process.run` (#16306)
The tested behaviours differ greatly between platforms. So there are lots of flag-based skips. And I'm not sure these specs work 100% correctly on all platforms.
Besides the typical CI machines, I've only tested it on FreeBSD.

We still need specs even for this platform-specific behaviour to ensure the aspects that we control in our system bindings are correct (e.g. #6464). 
The `$PATH` related specs are also a preparation for implementing a native fallback for `execvpe`.