Implement `execvpe_impl` (#16322)
Fills the stub `execvpe_impl` fallback for platforms without `LibC.execvpe` with a proper implementation.

The flag `-Dexecvpe_impl` enables this implementation even when libc provides this function. This has proven to be useful for testing purposes, and we might keep it around for a while (there will be more follow-up changes where it could be useful), maybe indefinitely.

Some of the design decisions for edge case behaviour were not exactly straightforward because system libraries handle things very differently sometimes. See https://github.com/crystal-lang/crystal/issues/6464#issuecomment-3455869962 for more details.

I usually opted for the most common and simple variant. We can refine details in the future if the need arises.

* default path is `/usr/bin:/bin` (like darwin and similar to mosts BSDs).
* too long pathnames are skipped without warning
* check executability by trying `execve`
* `NOEXEC` error does not try to execute a shell script.

The following system implementations served as role models:

* https://github.com/bminor/glibc/blob/eda354cfd8b998b55e4d62f1ad8aecf128be8d3c/posix/execvpe.c#L71
* https://github.com/bminor/musl/blob/1b76ff0767d01df72f692806ee5adee13c67ef88/src/process/execvp.c#L9
* https://github.com/jart/cosmopolitan/blob/f1e83d52403060d674161944e849b51f95707c9a/libc/proc/execvpe.c#L43
* https://developer.apple.com/library/archive/documentation/System/Conceptual/ManPages_iPhoneOS/man3/execvp.3.html
* https://github.com/freebsd/freebsd-src/blob/e5f545e0d872fbfdb6c3c94ca81e7f3bce9d094a/lib/libc/gen/exec.c#L148
* https://github.com/openbsd/src/blob/5ec89147ad5e4d03fe2ebfa9305d32a8b448119b/lib/libc/gen/exec.c#L184
* https://github.com/NetBSD/src/blob/0e31880b9b2f168b1e26bc9c096f5537f4f36a9d/lib/libc/gen/execvp.c#L59
* https://github.com/DragonFlyBSD/DragonFlyBSD/blob/4b5fc24f2f4611b38e1f3232524b57c45a5b3bb4/lib/libc/gen/exec.c#L145
* https://github.com/illumos/illumos-gate/blob/56ae62f22e43044050cd303d4ea63c1d556a8425/usr/src/lib/libc/port/gen/execvp.c#L109
* https://github.com/LineageOS/android_bionic/blob/5e704b240138776d1b6a54892d3619424ff4016a/libc/bionic/exec.cpp#L119

Co-authored-by: Julien Portalier <julien@portalier.com>