Add `Random#split` and `#split_internal` API for splittable PRNGs (#16342)
Adds a couple methods to formalize the API for splitting a `Random` instance:

- `Random#split`: splits the instance by allocating a new instance.
- `#split_internal`: splits the instance into an allocated instance (for example `ReferenceStorage(T)`).

I initially wanted a `#split!` method with no argument, but we need access to both instances to avoid splits (and re-splits of splits) to end up sharing the same sequences.

Extracted from #16174 and adapted with usages in the [random shard](https://github.com/ysbaddaden/random.cr/pull/1)(SplittableRandom, XoShiRo).