Add `Pointer#fill` (#16338)
This is just shifting the method down from `Slice#fill`.
The existing implementation already used almost exclusively pointer features, which serves as an indicator that the functionality belongs to `Pointer` with `Slice` just providing a convenient and bounds-checked wrapper.
It's also useful for `Pointer.malloc` which basically already implemented the behaviour, just without the optimizations. A couple other methods could be based on `#fill` as well.

There are already similar methods like the `#map` family in `Pointer`, so it should be expected that `#fill` fits well in the public API.

Ref https://github.com/crystal-lang/crystal/issues/16203#issuecomment-3472156133