Fix `Env.make_envp` add final null pointer (#16384)
The final entry in `envp` must be a null pointer. This would usually be the case implicitly because unused items in the array buffer are zeroed out. But in case we fill the array to capacity, there is no guarantee of a final null pointer. We need to make that explicit.