Build compiler with `-Dpreview_mt` (#16380)
We added support for multi-threaded codegen in #14748 but never enabled it in the default build configuration. It's expected to be more efficient than the current fork-based parallelization. And it works on Windows.

I suppose we could also consider passing `-Dexecution_context`? But the `mt_codegen` code probably needs an adjustment for that (at least resize the default context to `n_threads`.

Compiler versions before 1.8 are broken with `-Dpreview_mt`, so we need to disable it for older versions.