Fix reset directory if `Process.exec` fails (#16383)
When we try to exec a new process image, this might fail for a number of reason (e.g. executable not found or not executable). As a result, the current process continues and `Process.exec` raises an appropriate exception. This exception can be caught somewhere, and we should avoid any changes intended for the _new_ process image to permanently affect the current process' environment.
Using `Dir.cd` with a block ensures the directory is reset at the end of the block.