Add thread safety to ENV (breaking change)
Creates a snapshot of the system environment during program startup then
only reads from and writes to the snapshot. The environment is thus
local to the crystal program, but still behaves normally: the parent
environment is available to the process and local environment is
exported to the the child processes normally.

Adds `ENV.unsafe_get` and `ENV.unsafe_set` for the very edge case where
the program needs to access the system environment at runtime to
communicate with external libraries.