Ensure single reader and writer to system fd on Unix (#16209)
This patch extends the fdlock to **serialize reads and writes** by extending the reference counted lock with a read lock and a write lock, so taking a reference and locking acts as a single operation instead of two (1. acquire/release the lock; 2. take/return a reference). This avoids a race condition in the polling event loops:

- Fiber 1 then Fiber 2 try to read from `fd`;
- Since `fd` isn't ready, both fibers start waiting;
- When `fd` becomes ready then Fiber 1 is resumed;
- Fiber 1 doesn't read everything and _returns_;
- Since events are edge-triggered, Fiber 2 won't be resumed!!!

With the read lock, fiber 2 will wait on the lock then be resumed by fiber 1 when it returns. A concrete example is multiple fibers waiting to accept on a socket where fiber 1 would keep handling connections, while fiber 2 sits idle.

The other benefit is that it can help to simplify the evloops that will now only deal with a single reader + single writer per `IO` and is required for the io_uring evloop (the MT version requires it).

**NOTE**: While this patch only serializes reads/writes on UNIX at the `Crystal::System`, which is where the bugs are, we will move it into stdlib for all targets in a follow-up. See https://github.com/crystal-lang/crystal/pull/16289#issuecomment-3562748047