Resize execution context for spec with `CRYSTAL_WORKERS` (#16444)
When running specs with execution contexts, it's annoying to have to explicitly scale the default execution context in order to get actual parallelism. This could be handled in some kind of spec helper, but it would still require extra effort, and it'd be easy to miss out on it.

Thus we resize the EC in the stdlib `spec` package (i.e. for _any_ spec program) if the environment variable `CRYSTAL_WORKERS` is defined.