Build compiler with `-Dexecution_context` (#16447)
We just enabled MT in #16380, but we can go further and enable execution context as suggested in the PR, by merely resizing the default context to the default workers count.

Also changes the default `--threads` value to also be the default workers count, which is `CRYSTAL_WORKERS` when specified and valid, and defaults to the number of available logical CPUs instead of being hardcoded to 4 threads (default for `preview_mt`) or 8 forked processes (arbitrary number).

It might be faster, on my computer recompiling Crystal went from ~17s to ~15s but your mileage may vary.