LLVM: simplify target initialization and support more targets (#16437)
The LLVM bindings only expose the targets that the Crystal compiler itself supports, and each target copy-pastes the LibLLVM bindings and LLVM init methods which should use a macro.

I introduce a `LibLLVM::ALL_TARGETS` constant that's only used in macros to check if a known target has been built and generate the bindings and init methods accordingly.

The `@@initialized_` class variables are now an atomic, and the dead `@@initialized` has been removed.