Fix: execution context queue stress tests failures
Refactors and abstracts the stress test runs so no loop will run
forever: the "thread setup" part has been removed, the main fiber won't
block waiting for the threads to be ready, and the threads' loop will
eventually timeout, and the thread return, so the main fiber won't block
while joining the threads.

This fixes the regular CI failures that occured often on Darwin, and may
happen on Linux when running both specs in a tight loops multiple times
in parallel to overload the CPU cores.