Add `Process.pthread_disable_cancelstate` (#16446)
`fork` in a multi-threaded process is a very delicate matter. We must avoid cancelling the thread while performing a fork. Cleanup could cause problems while we're setting up the new process.

This would not be necessary if we were using `vfork` semantics. That's generally the better choice for spawning a new process, as there is no need to copy the entire process memory. `vfork` is more efficient, and we should ideally use that.

However, we're hoping to switch to `posix_spawn` eventually (#11337), which provides a safe and efficient implementation, and we don't need to deal with the details explicitly.
This currently blocked on the need to reopen stdio in the housekeeping stage (#16353) but we're expecting to resolve that soon.