Fix use after unlink in `XML::Node` (#16432)
This code works because attributes only have a single text node in practice, but the use after unlink is troublesome. The code also checks (`obj = cached?(node)`) the parent when iterating over the children. There is no memory leak here, but I am including the PR since I think it brings the implementation in line with the name of the method.