Fix execution context queue stress tests failures (#16472)
Refactors and abstracts the stress test specs of `Fiber::ExecutionContext::Runnables` and `Fiber::ExecutionContext::GlobalQueue` so no loop will run forever: the "thread setup" part has been removed, the main fiber won't block waiting for the threads to be ready, and the threads' loop will eventually timeout, and the thread return, so the main fiber won't block while joining the threads.

I abstracted a helper because the different tests used the same structure, and it was painful & noisy to dup the logic.

This fixes the regular CI failures that often occurred on Darwin on CI, and that I just reproduced on Linux when running these specs in tight loops multiple times in parallel to overload the CPU cores.

Co-authored-by: Johannes Müller <straightshoota@gmail.com>