Fix: interpreter musn't reuse dead fiber stack
The interpreter keeps an internal pool of stacks to interpret fibers and
C callbacks, and immediately releases the stack when it's done with it.

That pool instance musn't consider `Thread.current.dead_fiber_stack`
that's meant for the execution context scheduler where the interpreter
lives in (i.e. another pool).