Makefile: Use simply expanded variables to avoid costly duplicate evaluation (#16509)
We're using `?=` to assign variables only if they are not already defined. This is a recursively expanded variable assignment which is evaluated on every variable use. 
This meant for `LLVM_CONFIG`, the shell script might execute multiple time per make instance. But we only need it to run once and cache the result.
That's why we should use simply expanded variables (assigned with `:=`). But that means we cannot use the shorthand `?=` and need an explicit check if the variable is defined. This is a bit more boilerplate, unfortunately.