Fix: interpreter musn't reuse dead fiber stacks (#16518)
The interpreter keeps an internal pool of stacks to interpret fibers and C callbacks, and immediately releases the stack when it's done with it.

The pool instance mustn't consider `Thread.current.dead_fiber_stack` that is meant for the execution context where the interpreter lives (i.e. another pool), and both stacks can't be interchanged between pools: the scheduler stacks grow down, but the interpreter stacks grow up, and we protect the bottom of scheduler stacks and :boom:

I don't know why it only segfaults on Windows, though.